requirements.txt
gradio==6.0.1
requests==2.28.2
Pillow==12.0.0
python-dotenv==1.2.1
qrcode==8.2

# Notes:
# - These versions were taken from the current .venv in this workspace (gradio 6.0.1, requests 2.28.2, Pillow 12.0.0, python-dotenv 1.2.1, qrcode 8.2).
# - If you'd prefer looser version constraints use '>=' instead of '=='.